Moved Filter function to _filtering file
